requirements.txt
Django
python-dotenv
psycopg2-binary
whitenoise
djangorestframework
requests
cryptography
celery
redis
django-celery-beat
django-celery-results
django-cors-headers
djangorestframework-simplejwt
gunicorn
pytz
django-hijack
django-extensions